requirements.txt
pandas
requests
tqdm
beautifulsoup4
lxml
scikit-learn
qc-B3DB
joblib